Renamed AQHI xml parser.
